recipe_site/requirements.txt

Django>=4.0
gunicorn
python-dotenv
dj-database-url
psycopg2-binary
djangorestframework